// packages/core/src/db/schema.ts
export interface CategoryTable {
  id: string;
  name: string;
  type: TransactionType;
  color: string;
  icon: string;
  isDefault: SqlBool;
  userId: string;
  createdAt: string;
  updatedAt: string;
}